Update accounthing.sh with Bash-specific syntax
Changed `[]` to the improved `[[]]` which does not perform word
splitting or filename expansion (main reason to double quote
parameter expansions).

Used `-z` in line 131 to be consistent across the script.

Removed the parentheses in line 149 since there is no real reason
to execute the conditional statements in a subshell and the `[[]]`
allow for multiple statements within a single set of brackets.